requirements.txt
pyserial>=3.5
cryptography>=41.0.0 